Refactored ability and damage
